api-server-flask/requirements.txt
aniso8601==9.0.1
attrs==22.1.0
bcrypt==4.0.0
click==8.1.3
colorama==0.4.5
Flask==2.1.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.3
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.15.0
MarkupSafe==2.1.1
nose==1.3.7
PyJWT==2.4.0
pyrsistent==0.18.1
pytz==2022.2.1
six==1.16.0
SQLAlchemy==1.4.41
Werkzeug==2.1.2
